+++ packages/integration-sdk-entity-validator/src/index.ts
@@ -4,3 +4,4 @@ export {
   isEntityValidationError,
   type EntityValidationError,
 } from './entityValidationError';
+export { getValidator, getValidatorSync } from './singleton';