+++ CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to
 
 # Unreleased
 
+# 13.3.0 - 2024-07-16
+
+- Added a `toMatchDataModelSchema` Jest matcher for validating entities against
+  the live production data model schema.
+
 # 13.2.0 - 2024-07-11
 
 - De-nest resulting single-class validation schema.